# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  success: boolean;
  message: string;
  errorCode?: number;
}

export interface CreateTokenRequest {
  clientId: string;
  clientSecret: string;
  apiEndpoint?: string;
}

export interface CreateTokenResponse {
  success: boolean;
  token?: string;
  expiration?: number;
  message?: string;
}

export interface Vendor {
  vendorCode: string;
  /** 1=live casino, 2=slot, 3=mini-game */
  type: number;
  name: string;
  url?: string;
}

export interface VendorListResponse {
  success: boolean;
  message?: Vendor[];
  errorCode?: number;
}

export interface GetGamesRequest {
  vendorCode: string;
  language: string;
}

export interface Game {
  provider?: string;
  vendorCode: string;
  gameId?: string;
  gameCode: string;
  gameName: string;
  slug?: string;
  thumbnail?: string;
  updatedAt?: string;
  isNew?: boolean;
  underMaintenance?: boolean;
}

export interface GameListResponse {
  success: boolean;
  message?: Game[];
  errorCode?: number;
}

export interface GetGameDetailRequest {
  vendorCode: string;
  gameCode: string;
  language?: string;
}

export interface GameDetailResponse {
  success: boolean;
  message?: Game | string;
  errorCode?: number;
}

export interface LaunchGameRequest {
  vendorCode: string;
  gameCode: string;
  playerCode: string;
  language?: string;
  homeUrl?: string;
  depositUrl?: string;
}

export interface LaunchGameResponse {
  success: boolean;
  /** The game URL to load in iframe */
  message?: string;
  errorCode?: number;
}

export interface CreatePlayerRequest {
  playerCode: string;
  currency?: string;
  nickname?: string;
}

export interface PlayerResponse {
  success: boolean;
  message?: string;
  errorCode?: number;
}

export interface PlayerCodeRequest {
  playerCode: string;
}

export interface BalanceResponse {
  success: boolean;
  message?: number;
  errorCode?: number;
}

export interface DepositRequest {
  playerCode: string;
  amount: number;
  txnId?: string;
}

export interface WithdrawRequest {
  playerCode: string;
  amount: number;
  txnId?: string;
}

export interface ConfigResponse {
  hasCredentials: boolean;
  apiEndpoint?: string;
}