# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Creates an authentication token for OroPlay API
 * @summary Get OroPlay auth token
 */
export const CreateTokenBody = zod.object({
  clientId: zod.string(),
  clientSecret: zod.string(),
  apiEndpoint: zod.string().optional(),
});

export const CreateTokenResponse = zod.object({
  success: zod.boolean(),
  token: zod.string().optional(),
  expiration: zod.number().optional(),
  message: zod.string().optional(),
});

/**
 * @summary Get list of game vendors/providers
 */
export const GetVendorsResponse = zod.object({
  success: zod.boolean(),
  message: zod
    .array(
      zod.object({
        vendorCode: zod.string(),
        type: zod.number().describe("1=live casino, 2=slot, 3=mini-game"),
        name: zod.string(),
        url: zod.string().optional(),
      }),
    )
    .optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Get list of games for a vendor
 */
export const GetGamesBody = zod.object({
  vendorCode: zod.string(),
  language: zod.string(),
});

export const GetGamesResponse = zod.object({
  success: zod.boolean(),
  message: zod
    .array(
      zod.object({
        provider: zod.string().optional(),
        vendorCode: zod.string(),
        gameId: zod.string().optional(),
        gameCode: zod.string(),
        gameName: zod.string(),
        slug: zod.string().optional(),
        thumbnail: zod.string().optional(),
        updatedAt: zod.string().optional(),
        isNew: zod.boolean().optional(),
        underMaintenance: zod.boolean().optional(),
      }),
    )
    .optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Get game detail
 */
export const GetGameDetailBody = zod.object({
  vendorCode: zod.string(),
  gameCode: zod.string(),
  language: zod.string().optional(),
});

export const GetGameDetailResponse = zod.object({
  success: zod.boolean(),
  message: zod
    .union([
      zod.object({
        provider: zod.string().optional(),
        vendorCode: zod.string(),
        gameId: zod.string().optional(),
        gameCode: zod.string(),
        gameName: zod.string(),
        slug: zod.string().optional(),
        thumbnail: zod.string().optional(),
        updatedAt: zod.string().optional(),
        isNew: zod.boolean().optional(),
        underMaintenance: zod.boolean().optional(),
      }),
      zod.string(),
    ])
    .optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Launch a game and get the game URL
 */
export const LaunchGameBody = zod.object({
  vendorCode: zod.string(),
  gameCode: zod.string(),
  playerCode: zod.string(),
  language: zod.string().optional(),
  homeUrl: zod.string().optional(),
  depositUrl: zod.string().optional(),
});

export const LaunchGameResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional().describe("The game URL to load in iframe"),
  errorCode: zod.number().optional(),
});

/**
 * @summary Create or register a test player
 */
export const CreatePlayerBody = zod.object({
  playerCode: zod.string(),
  currency: zod.string().optional(),
  nickname: zod.string().optional(),
});

export const CreatePlayerResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Get player balance
 */
export const GetPlayerBalanceBody = zod.object({
  playerCode: zod.string(),
});

export const GetPlayerBalanceResponse = zod.object({
  success: zod.boolean(),
  message: zod.number().optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Deposit balance to player
 */
export const DepositBalanceBody = zod.object({
  playerCode: zod.string(),
  amount: zod.number(),
  txnId: zod.string().optional(),
});

export const DepositBalanceResponse = zod.object({
  success: zod.boolean(),
  message: zod.number().optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Withdraw balance from player
 */
export const WithdrawBalanceBody = zod.object({
  playerCode: zod.string(),
  amount: zod.number(),
  txnId: zod.string().optional(),
});

export const WithdrawBalanceResponse = zod.object({
  success: zod.boolean(),
  message: zod.number().optional(),
  errorCode: zod.number().optional(),
});

/**
 * @summary Get current OroPlay config (clientId present, no secrets)
 */
export const GetConfigResponse = zod.object({
  hasCredentials: zod.boolean(),
  apiEndpoint: zod.string().optional(),
});