# lib/api-zod/src/generated/types/createTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTokenResponse {
  success: boolean;
  token?: string;
  expiration?: number;
  message?: string;
}