# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  provider?: string;
  vendorCode: string;
  gameId?: string;
  gameCode: string;
  gameName: string;
  slug?: string;
  thumbnail?: string;
  updatedAt?: string;
  isNew?: boolean;
  underMaintenance?: boolean;
}