# lib/api-zod/src/generated/types/gameListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from "./game";

export interface GameListResponse {
  success: boolean;
  message?: Game[];
  errorCode?: number;
}