# lib/api-zod/src/generated/types/launchGameRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */

export interface LaunchGameRequest {
  vendorCode: string;
  gameCode: string;
  playerCode: string;
  language?: string;
  homeUrl?: string;
  depositUrl?: string;
}