# lib/api-zod/src/generated/types/launchGameResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */

export interface LaunchGameResponse {
  success: boolean;
  /** The game URL to load in iframe */
  message?: string;
  errorCode?: number;
}