# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */

export interface Vendor {
  vendorCode: string;
  /** 1=live casino, 2=slot, 3=mini-game */
  type: number;
  name: string;
  url?: string;
}