# lib/api-zod/src/generated/types/vendorListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casino Lobby API - OroPlay proxy
 * OpenAPI spec version: 0.1.0
 */
import type { Vendor } from "./vendor";

export interface VendorListResponse {
  success: boolean;
  message?: Vendor[];
  errorCode?: number;
}